refactor some units
