+++ project/src/components/film-card-details/simple-text.tsx
